+++ CHANGES.rst
@@ -1,7 +1,7 @@
 Upcoming
 ========
     * The unpickler is now more resilient to malformed "py/reduce", "py/set",
-      "py/tuple", and "py/iterator" input data. (+544)
+      "py/tuple", "py/b64", "py/b85", and "py/iterator" input data. (+544) (+545)
     * The test suite was updated to leverage more pytest features.
     * The ``jsonpickle.compat`` module is no longer used. It is still provided
       for backwards compatibility but it may be removed in a future version.